@@ -67,31 +66,32 @@
         "symfony/twig-bridge": "6.0.*",
         "symfony/twig-bundle": "6.0.*",
         "symfony/twilio-notifier": "6.0.*",
-        "symfony/ux-autocomplete": "^2.2",
+        "symfony/ux-autocomplete": "^2.4",
         "symfony/ux-chartjs": "^2.4",
-        "symfony/ux-cropperjs": "^2.2",
+        "symfony/ux-cropperjs": "^2.4",
         "symfony/ux-dropzone": "^2.2",
-        "symfony/ux-lazy-image": "^2.2",
-        "symfony/ux-live-component": "^2.2",
-        "symfony/ux-notify": "^2.2",
-        "symfony/ux-react": "^2.2",
-        "symfony/ux-swup": "^2.2",
-        "symfony/ux-turbo": "^2.2",
-        "symfony/ux-twig-component": "^2.2",
-        "symfony/ux-typed": "^2.2",
+        "symfony/ux-lazy-image": "^2.4",
+        "symfony/ux-live-component": "^2.4",
+        "symfony/ux-notify": "^2.4",
+        "symfony/ux-react": "^2.4",
+        "symfony/ux-swup": "^2.4",
+        "symfony/ux-turbo": "^2.4",
+        "symfony/ux-twig-component": "^2.4",
+        "symfony/ux-typed": "^2.4",
         "symfony/validator": "6.0.*",
-        "symfony/webpack-encore-bundle": "^1.14",
+        "symfony/webpack-encore-bundle": "^1.15",
         "symfony/yaml": "6.0.*",
-        "symfonycasts/verify-email-bundle": "^v1.7.0",
-        "twig/extra-bundle": "^3.3.0",
-        "twig/intl-extra": "^3.3",
-        "twig/string-extra": "^3.3",
-        "twig/twig": "^3.3.0",
-        "twilio/sdk": "^6.31",
-        "vich/uploader-bundle": "^1.21.1"
+        "symfonycasts/verify-email-bundle": "^1.11.0",
+        "twig/extra-bundle": "^3.4.0",
+        "twig/intl-extra": "^3.4",
+        "twig/string-extra": "^3.4",
+        "twig/twig": "^3.4.0",
+        "twilio/sdk": "^6.42",
+        "vich/uploader-bundle": "^1.21.1",
+        "webmozart/assert": "1.11.*"
     },
     "require-dev": {
-        "dama/doctrine-test-bundle": "^6",
+        "dama/doctrine-test-bundle": "^6.7",
         "doctrine/doctrine-fixtures-bundle": "^3.4",
         "fzaninotto/faker": "^1.5",
         "phpstan/phpstan-symfony": "^1.2.5",
@@ -100,8 +100,8 @@
         "symfony/browser-kit": "6.0.*",
         "symfony/css-selector": "6.0.*",
         "symfony/debug-bundle": "6.0.*",
-        "symfony/maker-bundle": "^1.43",
-        "symfony/monolog-bundle": "^3.0",
+        "symfony/maker-bundle": "^1.45",
+        "symfony/monolog-bundle": "^3.8",
         "symfony/panther": "^2.0",
         "symfony/phpunit-bridge": "^6.1",
         "symfony/stopwatch": "6.0.*",
