@@ -5,9 +5,11 @@ namespace App\Controller\Admin;
 
 
 use App\Entity\Category\CategoryAttachment;
+use App\Form\Attachment\AttachmentType;
 use EasyCorp\Bundle\EasyAdminBundle\Config\Crud;
 use EasyCorp\Bundle\EasyAdminBundle\Controller\AbstractCrudController;
 use EasyCorp\Bundle\EasyAdminBundle\Field\AssociationField;
+use EasyCorp\Bundle\EasyAdminBundle\Field\CollectionField;
 use EasyCorp\Bundle\EasyAdminBundle\Field\ImageField;
 use EasyCorp\Bundle\EasyAdminBundle\Field\TextEditorField;
 use EasyCorp\Bundle\EasyAdminBundle\Field\TextField;
@@ -22,26 +24,26 @@ class CategoryAttachmentCrudController extends AbstractCrudController
 
     public function configureFields(string $pageName): iterable
     {
-
-        $thumbnailFile = ImageField::new('thumbnailFile')->setFormType(VichImageType::class);
-        $thumbnail = ImageField::new('fileVich')->setBasePath('/upload/category/thumbnail');
-        $fields = [
-            TextField::new('first_title'),
+        return [
+            TextField::new('firstTitle'),
             TextEditorField::new('middle_title'),
             TextField::new('last_title')->setMaxLength(48)->onlyOnIndex(),
             TextEditorField::new('last_title')->setNumOfRows(10)->hideOnIndex(),
-            AssociationField::new('categoryAttachmentCategory')->autocomplete(), //TODO: доработать контракт
-            TextField::new('create_By')->hideOnForm(),//TODO: доработать контракт
-
+            CollectionField::new('categoryAttachment')
+                ->setEntryType(AttachmentType::class)
+                ->setFormTypeOption('by_reference', false)
+                ->onlyOnForms()
+            ,
+            CollectionField::new('categoryAttachment')
+                ->setTemplatePath('images.html.twig')
+                ->onlyOnDetail()
+            ,
+//            ImageField::new('firstTitle')
+//                ->setBasePath('/upload/category/thumbnail')
+//                ->setUploadDir('/upload/category/thumbnail')
+//                ->onlyOnIndex()
+//            ,
         ];
-
-        ($pageName == Crud::PAGE_INDEX || $pageName == Crud::PAGE_DETAIL) ?
-            $fields[] = $thumbnail :
-            $fields[] = $thumbnailFile;
-
-        return $fields;
-
-
     }
 
 
