@@ -2,19 +2,21 @@
 
 namespace App\Controller\Admin;
 
+use App\Entity\Attachment\Attachment;
 use App\Entity\Category\Category;
 use App\Entity\Category\CategoryAttachment;
 use App\Entity\Product\Product;
+use App\Entity\Product\ProductAttachment;
 use App\Entity\Product\ProductPrice;
 use App\Entity\Product\ProductTag;
 use App\Entity\Project\Project;
 use App\Entity\Project\ProjectAttachment;
+use App\Entity\Project\ProjectPlatformReward;
 use App\Entity\Project\ProjectTag;
-use App\Entity\Review\ReviewProduct;
-use App\Entity\Review\ReviewProject;
+use App\Entity\Review\ProductReview;
+use App\Entity\Review\ProjectReview;
 use App\Entity\Vendor\Vendor;
 use App\Entity\Vendor\VendorDocument;
-use App\Entity\Vendor\VendorEnGb;
 use App\Entity\Vendor\VendorMedia;
 use EasyCorp\Bundle\EasyAdminBundle\Config\Dashboard;
 use EasyCorp\Bundle\EasyAdminBundle\Config\MenuItem;
@@ -23,9 +25,10 @@ use EasyCorp\Bundle\EasyAdminBundle\Controller\AbstractDashboardController;
 use EasyCorp\Bundle\EasyAdminBundle\Router\AdminUrlGenerator;
 use Psr\Container\ContainerExceptionInterface;
 use Psr\Container\NotFoundExceptionInterface;
-use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
+use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Security\Core\User\UserInterface;
+
 
 
 class DashboardController extends AbstractDashboardController
@@ -42,11 +45,14 @@ class DashboardController extends AbstractDashboardController
      * @throws ContainerExceptionInterface
      * @throws NotFoundExceptionInterface
      */
-    #[Route(path: '/easyadmin', name: 'easyadmin')]
-    public function index() : Response
+    #[Route(path: '/easyadmin', name: 'easyadmin_index')]
+    public function index(): Response
     {
-        $routeBuilder = $this->container->get(AdminUrlGenerator::class);
-        return $this->redirect($routeBuilder->setController(CategoryEnGbCrudController::class)->generateUrl());
+        $adminUrlGenerator = $this->container->get(AdminUrlGenerator::class);
+
+        // Option 1. Make your dashboard redirect to the same page for all users
+        return $this->redirect($adminUrlGenerator->setController(ProjectCrudController::class)->generateUrl());
+
     }
 
     public function configureDashboard(): Dashboard
@@ -59,48 +65,48 @@ class DashboardController extends AbstractDashboardController
             ->setTextDirection('ltr')
             ->renderContentMaximized()
             # ->renderSidebarMinimized()
-            ->disableUrlSignatures()
             ->generateRelativeUrls()
+            ->setTranslationDomain('message')
             ;
     }
+
+
 
     public function configureMenuItems(): iterable
     {
         yield MenuItem::linktoDashboard('Dashboard', 'fa fa-home');
-
-        yield MenuItem::section('Projects');
+        #
+        yield MenuItem::section('Project');
         yield MenuItem::linkToCrud('Projects', 'fa fa-tags', Project::class);
-//        yield MenuItem::section('ProjectAttachments');
-        yield MenuItem::linkToCrud('Projects', 'fa fa-tags', ProjectAttachment::class);
-//        yield MenuItem::section('Rewards');
-//        yield MenuItem::linkToCrud('Rewards', 'fa fa-tags', ProjectPlatformReward::class);
-
+        yield MenuItem::linkToCrud('PlatformReward', 'fa fa-tags', ProjectPlatformReward::class);
+        yield MenuItem::linkToCrud('ProjectMedia', 'fa fa-tags', ProjectAttachment::class);
+        yield MenuItem::linkToCrud('ProjectReview', 'fa fa-tags', ProjectReview::class);
+        #
         yield MenuItem::section('Product');
-        yield MenuItem::linkToCrud('Product', 'fa fa-tags', Product::class);
-//        yield MenuItem::section('Price');
-        yield MenuItem::linkToCrud('Price', 'fa fa-tags', ProductPrice::class);
-
-        yield MenuItem::section('Users|Vendors');
+        yield MenuItem::linkToCrud('Products', 'fa fa-tags', Product::class);
+        yield MenuItem::linkToCrud('ProductMedia', 'fa fa-tags', ProductAttachment::class);
+        yield MenuItem::linkToCrud('ProductPrice', 'fa fa-tags', ProductPrice::class);
+        yield MenuItem::linkToCrud('ProductReview', 'fa fa-tags', ProductReview::class);
+        #
+        yield MenuItem::section('User|Vendor');
         yield MenuItem::linkToCrud('Vendors', 'fa fa-tags', Vendor::class);
-//        yield MenuItem::section('VendorAttachments');
-        yield MenuItem::linkToCrud('VendorDocument', 'fa fa-tags', VendorDocument::class);
         yield MenuItem::linkToCrud('VendorMedia', 'fa fa-tags', VendorMedia::class);
-
-        yield MenuItem::section('Categories');
+        yield MenuItem::linkToCrud('VendorDocument', 'fa fa-tags', VendorDocument::class);
+        #
+        yield MenuItem::section('Category');
         yield MenuItem::linkToCrud('Categories', 'fa fa-tags', Category::class);
-//        yield MenuItem::section('CategoryAttachments');
-        yield MenuItem::linkToCrud('CategoryAttachment', 'fa fa-tags', CategoryAttachment::class);
-
-        yield MenuItem::section('Reviews');
-        yield MenuItem::linkToCrud('ProductReviews', 'fa fa-tags', ReviewProduct::class);
-        yield MenuItem::linkToCrud('ProjectReviews', 'fa fa-tags', ReviewProject::class);
-
+        yield MenuItem::linkToCrud('CategoryMedia', 'fa fa-tags', CategoryAttachment::class);
+        #
+        yield MenuItem::section('Review');
+        yield MenuItem::linkToCrud('ProductReview', 'fa fa-tags', ProductReview::class);
+        yield MenuItem::linkToCrud('ProjectReview', 'fa fa-tags', ProjectReview::class);
+        #
         yield MenuItem::section('Tags');
         yield MenuItem::linkToCrud('ProjectTags', 'fa fa-tags', ProjectTag::class);
         yield MenuItem::linkToCrud('ProductTags', 'fa fa-tags', ProductTag::class);
-
+        #
         yield MenuItem::section('Events');
-
+        #
         yield MenuItem::linkToLogout('Logout', 'fa fa-exit');
         yield MenuItem::linkToExitImpersonation('Stop impersonation', 'fa fa-exit');
 
@@ -110,6 +116,7 @@ class DashboardController extends AbstractDashboardController
     {
         return parent::configureUserMenu($vendor)
             ->setName($vendor->getUserIdentifier());
+
 //        // use the given $user object to get the user name
 //        ->setName($user->getFullName())
 //        // use this method if you don't want to display the name of the user
