@@ -4,7 +4,7 @@
 namespace App\Controller\Admin;
 
 
-use App\Entity\Project\ProjectAttachment;
+use App\Entity\Project\ProductAttachment;
 use EasyCorp\Bundle\EasyAdminBundle\Controller\AbstractCrudController;
 use EasyCorp\Bundle\EasyAdminBundle\Field\IdField;
 use EasyCorp\Bundle\EasyAdminBundle\Field\TextEditorField;
@@ -14,9 +14,23 @@ class ProjectAttachmentCrudController extends AbstractCrudController
 {
     public static function getEntityFqcn(): string
     {
-        return ProjectAttachment::class;
+        return ProductAttachment::class;
     }
 
-    use ConfigureCRUDsFieldTrait;
+    public function configureFields(string $pageName): iterable
+    {
+
+        return [
+            TextField::new('firstTitle'),
+            TextEditorField::new('middle_title'),
+
+            TextEditorField::new('last_title'),
+            TextEditorField::new('last_title')->setNumOfRows(10)->hideOnIndex(),
+            TextField::new('last_title')->setMaxLength(48)->onlyOnIndex(),
+        ];
+
+
+    }
+
     use ConfigureFiltersTrait;
 }
