@@ -5,9 +5,12 @@ namespace App\Controller\Admin;
 
 
 use App\Entity\Project\Project;
+use App\Form\Project\ProjectAttachmentType;
+use App\Form\Vendor\VendorMediaType;
 use EasyCorp\Bundle\EasyAdminBundle\Config\Crud;
 use EasyCorp\Bundle\EasyAdminBundle\Controller\AbstractCrudController;
 use EasyCorp\Bundle\EasyAdminBundle\Field\AssociationField;
+use EasyCorp\Bundle\EasyAdminBundle\Field\CollectionField;
 use EasyCorp\Bundle\EasyAdminBundle\Field\ImageField;
 use EasyCorp\Bundle\EasyAdminBundle\Field\TextEditorField;
 use EasyCorp\Bundle\EasyAdminBundle\Field\TextField;
@@ -22,25 +25,27 @@ class ProjectCrudController extends AbstractCrudController
 
     public function configureFields(string $pageName): iterable
     {
-        $thumbnailFile = ImageField::new('thumbnailFile')->setFormType(VichImageType::class);
-        $thumbnail = ImageField::new('fileVich')->setBasePath('/upload/project/image/thumbnail');
-        $fields = [
-            TextField::new('first_title'),
+
+        return [
+            TextField::new('firstTitle'),
             TextEditorField::new('middle_title'),
             TextField::new('last_title')->setMaxLength(48)->onlyOnIndex(),
             TextEditorField::new('last_title')->setNumOfRows(10)->hideOnIndex(),
-            AssociationField::new('projectProduct')->autocomplete()->hideOnIndex(),
-            TextField::new('create_By')->hideOnForm(),
-            AssociationField::new('projectAttachment')->hideOnIndex()->autocomplete(),
-            AssociationField::new('projectTag')->hideOnIndex()->autocomplete(),
-            AssociationField::new('projectPlatformReward')->hideOnIndex()->autocomplete(),
+            CollectionField::new('projectAttachment')
+                ->setEntryType(ProjectAttachmentType::class)
+                ->setFormTypeOption('by_reference', false)
+                ->onlyOnForms()
+            ,
+            CollectionField::new('projectAttachment')
+                ->setTemplatePath('images.html.twig')
+                ->onlyOnDetail()
+            ,
+//            ImageField::new('firstTitle')
+//                ->setBasePath('/upload/project/thumbnail')
+//                ->setUploadDir('/upload/project/thumbnail')
+//                ->onlyOnIndex()
+//            ,
         ];
-
-        ($pageName == Crud::PAGE_INDEX || $pageName == Crud::PAGE_DETAIL) ?
-            $fields[] = $thumbnail :
-            $fields[] = $thumbnailFile;
-
-        return $fields;
     }
 
     use ConfigureFiltersTrait;
