@@ -5,9 +5,11 @@ namespace App\Controller\Admin;
 
 
 use App\Entity\Vendor\VendorMedia;
+use App\Form\Vendor\VendorMediaType;
 use EasyCorp\Bundle\EasyAdminBundle\Config\Crud;
 use EasyCorp\Bundle\EasyAdminBundle\Controller\AbstractCrudController;
 use EasyCorp\Bundle\EasyAdminBundle\Field\AssociationField;
+use EasyCorp\Bundle\EasyAdminBundle\Field\CollectionField;
 use EasyCorp\Bundle\EasyAdminBundle\Field\ImageField;
 use EasyCorp\Bundle\EasyAdminBundle\Field\TextEditorField;
 use EasyCorp\Bundle\EasyAdminBundle\Field\TextField;
@@ -25,24 +27,42 @@ class VendorMediaCrudController extends AbstractCrudController
     {
 
 
-        $thumbnailFile = ImageField::new('thumbnailFile')->setFormType(VichImageType::class);
-        $thumbnail = ImageField::new('thumbnail')->setBasePath('/upload/vendor/image/thumbnail');
+//        $thumbnailFile = ImageField::new('thumbnailFile')->setFormType(VichImageType::class);
+//        $thumbnail = ImageField::new('thumbnail')->setBasePath('/upload/vendor/image/thumbnail');
 
-        $fields = [
+        //        ($pageName == Crud::PAGE_INDEX || $pageName == Crud::PAGE_DETAIL) ?
+//            $fields[] = $thumbnail :
+//            $fields[] = $thumbnailFile;
+
+        return [
             TextField::new('first_title'),
             TextEditorField::new('middle_title'),
+
+            TextEditorField::new('last_title'),
             TextField::new('last_title')->setMaxLength(48)->onlyOnIndex(),
             TextEditorField::new('last_title')->setNumOfRows(10)->hideOnIndex(),
-            AssociationField::new('vendorMedia')->autocomplete()->hideOnIndex(),
+
+            AssociationField::new('vendorMediaVendor')->autocomplete()->hideOnIndex(),
+
             TextField::new('create_By')->hideOnForm(),
-            AssociationField::new('vendorDocument')->hideOnIndex()->autocomplete(),
+
+            AssociationField::new('vendorMediaAttachment')->hideOnIndex()->autocomplete(),
+
+            CollectionField::new('vendorMediaAttachment')
+                ->setEntryType(VendorMediaType::class)
+                ->setFormTypeOption('by_reference', false)
+                ->onlyOnForms()
+            ,
+            CollectionField::new('vendorMediaAttachment')
+                ->setTemplatePath('admin/images.html.twig')
+                ->onlyOnDetail()
+            ,
+//            ImageField::new('fileName')
+//                ->setBasePath('/upload/vendor/thumbnail')
+//                ->setUploadDir('/upload/vendor/thumbnail')
+//                ->onlyOnIndex()
+//            ,
         ];
-
-        ($pageName == Crud::PAGE_INDEX || $pageName == Crud::PAGE_DETAIL) ?
-            $fields[] = $thumbnail :
-            $fields[] = $thumbnailFile;
-
-        return $fields;
     }
 
 
