@@ -3,7 +3,7 @@
 
 namespace App\Event;
 
-use App\Entity\Order\Order;
+use App\Entity\Order\OrderStorage;
 use Symfony\Contracts\EventDispatcher\Event;
 
 class OrderSubmitEvent extends Event
@@ -11,16 +11,16 @@ class OrderSubmitEvent extends Event
     public const NAME = 'orders.order';
 
     /**
-     * @param Order $orderSubmited
+     * @param OrderStorage $orderSubmited
      */
-    public function __construct(private readonly Order $orderSubmited)
+    public function __construct(private readonly OrderStorage $orderSubmited)
     {
     }
 
     /**
-     * @return Order
+     * @return OrderStorage
      */
-    public function getOrderSubmited(): Order
+    public function getOrderSubmited(): OrderStorage
     {
         return $this->orderSubmited;
     }
