@@ -3,7 +3,7 @@
 
 	namespace App\EventSubscriber;
 
-	use App\Entity\Order\Orders;
+	use App\Entity\Order\OrderStorage;
     use App\Event\Events;
     use App\Event\OrderSubmitEvent;
     use JetBrains\PhpStorm\ArrayShape;
@@ -13,49 +13,51 @@
     use Symfony\Component\EventDispatcher\EventSubscriberInterface;
     use Symfony\Component\EventDispatcher\GenericEvent;
     use Symfony\Component\Form\FormFactoryInterface;
+    use Symfony\Component\Mailer\Exception\TransportExceptionInterface;
     use Symfony\Component\Mailer\MailerInterface;
 
 	class OrderSubscriber implements EventSubscriberInterface
 	{
         #[NoReturn]
-		public function __construct(private MailerInterface $mailer, private ParameterBagInterface $params, private FormFactoryInterface $formFactory)
+		public function __construct(private readonly MailerInterface $mailer, private readonly ParameterBagInterface $params, private readonly FormFactoryInterface $formFactory)
 		{
 		}
 
 		#[ArrayShape([OrderSubmitEvent::NAME => "string"])]
         public static function getSubscribedEvents(): array
 		{
-			return array(
+			return [
 				OrderSubmitEvent::NAME => 'onOrdersOrder',
-                Events::ORDER_BEFORE_CREATE => 'onOrderBeforeCreate', //onPrePersist
-                Events::ORDER_CREATED => 'onOrderCreated', //onPostPersist
-                Events::ORDER_STATUS_UPDATED => 'onOrderStatusUpdated' // onUpdate
-			);
+//                Events::ORDER_BEFORE_CREATE => 'onOrderBeforeCreate', //onPrePersist
+//                Events::ORDER_CREATED => 'onOrderCreated', //onPostPersist
+//                Events::ORDER_STATUS_UPDATED => 'onOrderStatusUpdated' // onUpdate
+            ];
 		}
         //TODO: проработать нейминг методов
         #[Pure]
         public function onOrderBeforeCreate(GenericEvent $event)
         {
-            /** @var Orders $order */
+            /** @var OrderStorage $order */
             $order = $event->getSubject();
         }
 
         #[Pure]
         public function onOrderCreated(GenericEvent $event)
         {
-            /** @var Orders $order */
+            /** @var OrderStorage $order */
             $order = $event->getSubject();
         }
 
         #[Pure]
         public function onOrderStatusUpdated(GenericEvent $event)
         {
-            /** @var Orders $order */
+            /** @var OrderStorage $order */
             $order = $event->getSubject();
         }
 
-		public function onOrdersOrder(OrderSubmitEvent $orderSubmittedEvent): void
+
+        public function onOrdersOrder(OrderSubmitEvent $orderSubmittedEvent): void
 		{
-			$this->mailer->sendNewOrderNotification((array)$orderSubmittedEvent->getOrderSubmited());
+//			$this->mailer->send((array)$orderSubmittedEvent->getOrderSubmited()); //TODO: доработать мейл-нотификацию
 		}
 	}
