@@ -17,7 +17,7 @@ class ForgotPasswordType extends AbstractType
         $builder
             ->add('email', EmailType::class, [
                 'invalid_message' => 'The email address is invalid.',
-                'label'           => 'vendor.label.email',
+                'label'           => 'forgot.email.label',
                 'label_attr'      => [
                     'class' => 'sr-only',
                     'value' => 'last_username'
@@ -37,7 +37,7 @@ class ForgotPasswordType extends AbstractType
                     'id'          => 'email',
                     'name'        => '_email',
                     'class'       => '',
-                    'placeholder' => 'vendor.placeholder.email',
+                    'placeholder' => 'forgot.email.placeholder',
                     'tabindex'    => '101',
                     'required'    => null
                 ]
@@ -50,7 +50,7 @@ class ForgotPasswordType extends AbstractType
     {
         $resolver->setDefaults([
             'data_class'         => VendorSecurity::class,
-            'translation_domain' => 'security',
+            'translation_domain' => 'forgot',
             'validation_groups' => false,
             'attr' => [
                 'class' => 'needs-validation',
