@@ -3,7 +3,7 @@
 
 namespace App\Form\Order;
 
-use App\Entity\Order\Order;
+use App\Entity\Order\OrderStorage;
 use Symfony\Component\Form\AbstractType;
 use Symfony\Component\Form\FormBuilderInterface;
 use Symfony\Component\OptionsResolver\OptionsResolver;
@@ -61,7 +61,7 @@ class OrderType extends AbstractType
     public function configureOptions(OptionsResolver $resolver):void
     {
         $resolver->setDefaults([
-            'data_class' => Order::class,
+            'data_class' => OrderStorage::class,
         ]);
     }
 }
