@@ -12,6 +12,7 @@
 namespace App\Form\Type;
 
 use App\Utils\MomentFormatConverter;
+use Locale;
 use Symfony\Component\Form\AbstractType;
 use Symfony\Component\Form\Extension\Core\Type\DateTimeType;
 use Symfony\Component\Form\FormInterface;
@@ -28,7 +29,7 @@ use Symfony\Component\OptionsResolver\OptionsResolver;
  */
 class DateTimePickerType extends AbstractType
 {
-    public function __construct(private MomentFormatConverter $formatConverter)
+    public function __construct(private readonly MomentFormatConverter $formatConverter)
     {
     }
 
@@ -38,7 +39,7 @@ class DateTimePickerType extends AbstractType
     public function buildView(FormView $view, FormInterface $form, array $options): void
     {
         $view->vars['attr']['data-date-format'] = $this->formatConverter->convert($options['format']);
-        $view->vars['attr']['data-date-locale'] = mb_strtolower(str_replace('_', '-', \Locale::getDefault()));
+        $view->vars['attr']['data-date-locale'] = mb_strtolower(str_replace('_', '-', Locale::getDefault()));
     }
 
     /**
