@@ -3,22 +3,22 @@
 
 namespace App\Repository\Project;
 
-use App\Entity\Project\ProjectType;
+use App\Entity\Project\Type;
 use Doctrine\Bundle\DoctrineBundle\Repository\ServiceEntityRepository;
 use Doctrine\Persistence\ManagerRegistry;
 use Symfony\Component\HttpFoundation\InputBag;
 
 /**
- * @method ProjectType|null find($id, $lockMode = null, $lockVersion = null)
- * @method ProjectType|null findOneBy(array $criteria, array $orderBy = null)
- * @method ProjectType[]    findAll()
- * @method ProjectType[]    findBy(array $criteria, array $orderBy = null, $limit = null, $offset = null)
+ * @method Type|null find($id, $lockMode = null, $lockVersion = null)
+ * @method Type|null findOneBy(array $criteria, array $orderBy = null)
+ * @method Type[]    findAll()
+ * @method Type[]    findBy(array $criteria, array $orderBy = null, $limit = null, $offset = null)
  */
 class ProjectTypeRepository extends ServiceEntityRepository
 {
     public function __construct(ManagerRegistry $registry)
     {
-        parent::__construct($registry, ProjectType::class);
+        parent::__construct($registry, Type::class);
     }
 
     public function findBySearchQuery(float|InputBag|bool|int|string|null $query, float|InputBag|bool|int|string|null $limit)
